AboutPtypes

+++ koans/src/beginner/AboutPrimitives.java
@@ -67,8 +67,8 @@ public class AboutPrimitives {
 
     @Koan
     public void shortsHaveASmallerRangeThanInts() {
-        assertEquals(Short.MIN_VALUE, -32768);  // hint: You'll need an explicit cast
-        assertEquals(Short.MAX_VALUE, 32767);
+        assertEquals(Short.MIN_VALUE, (short) -32768);  // hint: You'll need an explicit cast
+        assertEquals(Short.MAX_VALUE, (short)32767);
     }
 
     @Koan
